pyproject.toml
[project]
name = "ai-server"
version = "0.1.0"
description = "Personal assistant server for Mac Mini: Telegram + web + cron → Claude Agent SDK."
requires-python = ">=3.12"
dependencies = [
    # Claude Agent SDK — the brain
    "claude-agent-sdk>=0.1.60",

    # Database
    "sqlalchemy[asyncio]>=2.0.29",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",

    # Redis (queue + pubsub)
    "redis>=5.0.3",

    # Web gateway
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.1",

    # Telegram gateway
    "python-telegram-bot>=21.0",

    # Scheduler
    "croniter>=2.0.5",

    # Manifest parsing
    "pyyaml>=6.0.1",

    # Logging
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.1",
    "pytest-asyncio>=0.23.6",
    "ruff>=0.4.0",
    "fakeredis>=2.22.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
